src/test/java/fr/t12/adventofcode/days/Day06Test.java
package fr.t12.adventofcode.days;

import org.junit.jupiter.api.Test;

import java.util.List;

import static org.junit.jupiter.api.Assertions.assertEquals;

public class Day06Test {

    private final Day06 day = new Day06();
    private final List<Integer> sampleInput = List.of(3, 4, 3, 1, 2);
    private final List<Integer> dayInput = day.getInput();

    @Test
    public void testResolvePart1WithSampleInput() {
        assertEquals(5_934L, day.resolvePart1(sampleInput));
    }

    @Test
    public void testResolvePart1WithDayInput() {
        assertEquals(362_346L, day.resolvePart1(dayInput));
    }

    @Test
    public void testResolvePart2WithSampleInput() {
        assertEquals(26_984_457_539L, day.resolvePart2(sampleInput));
    }

    @Test
    public void testResolvePart2WithDayInput() {
        assertEquals(1_639_643_057_051L, day.resolvePart2(dayInput));
    }
}